+++ CHANGES.md
@@ -8,6 +8,7 @@
 - Tests: Add test for `main` function
 - Tests: Run test harness on GHA
 - Tests: Add Windows to test matrix on CI/GHA
+- Tests: Fix tests on Windows
 
 
 ## 0.3.0 (2021-07-01)